Add msg_processor module and moved random_cob_id, random_msg and create_frame_send_msg from main.rs into the new module.
